+++ lib/widgets/time_value_row.dart
@@ -3,6 +3,7 @@ import 'package:timezone/src/date_time.dart';
 import 'package:timezone/timezone.dart' as tz;
 import 'package:url_launcher/url_launcher.dart';
 import '../l10n/app_localizations.dart';
+import '../layout_constants.dart';
 import '../main.dart';
 import '../models/time_value.dart';
 import '../models/timezone_abbr_localization.dart';